Stored the Coordinates of the generated obstacles and blocked collision between player and obstacle
Made a new HashSet called "BannedCoordinates" that will store the positions of the obstacles. In the random object position generator, all new generated coordinates are saved in the HashSet. The movement system has been updated to block the collision between the player and the obstacles by replacing the simple "RenderPlayer" with a position checker.