Made improvements
Renamed the methods to use the PascalCase text variant. Remade the "CalculateNewCoordinates" method to be an Extension Method. Also created a new "WithinBorders" extension method that checks whether the inputted coordinates are going to be within the playground borders and if they collide with an object. I also made a new "Playground" class that will contain the Console height, width and system rows.